+++ packages/vtable/__tests__/options/listTable-autoRowHeight.test.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck
 // 有问题可对照demo unitTestListTable
 import records from '../data/marketsales.json';
-import { ListTable } from '../../src/ListTable';
+import { ListTable } from '../../src';
 import { createDiv, removeDom } from '../dom';
 global.__VERSION__ = 'none';
 describe('listTable-autoRowHeight init test', () => {