+++ packages/vtable/src/event/listener/container-dom.ts
@@ -6,7 +6,8 @@ import { handleWhell } from '../scroll';
 import { browser } from '../../tools/helper';
 import type { EventManager } from '../event';
 import { getPixelRatio } from '../../tools/pixel-ratio';
-
+import { endResizeCol, endResizeRow } from './table-group';
+import { isCellDisableSelect } from '../../state/select/is-cell-select-highlight';
 export function bindContainerDomListener(eventManager: EventManager) {
   const table = eventManager.table;
   const stateManager = table.stateManager;
@@ -24,7 +25,9 @@ export function bindContainerDomListener(eventManager: EventManager) {
   });
 
   handler.on(table.getElement(), 'wheel', (e: WheelEvent) => {
-    handleWhell(e, stateManager);
+    if (table.eventManager._enableTableScroll) {
+      handleWhell(e, stateManager);
+    }
   });
 
   // 监听键盘事件
@@ -104,6 +107,10 @@ export function bindContainerDomListener(eventManager: EventManager) {
           targetCol = Math.min(table.colCount - 1, Math.max(0, stateManager.select.cellPos.col + 1));
         }
       }
+      // 如果是不支持选中的单元格 则退出
+      if (isCellDisableSelect(table, targetCol, targetRow)) {
+        return;
+      }
       table.selectCell(targetCol, targetRow, e.shiftKey);
       if (
         (table.options.keyboardOptions?.moveEditCellOnArrowKeys ?? false) &&
@@ -158,6 +165,10 @@ export function bindContainerDomListener(eventManager: EventManager) {
           } else {
             targetRow = stateManager.select.cellPos.row;
             targetCol = stateManager.select.cellPos.col + 1;
+          }
+          // 如果是不支持选中的单元格 则退出
+          if (isCellDisableSelect(table, targetCol, targetRow)) {
+            return;
           }
           table.selectCell(targetCol, targetRow);
           if ((table as ListTableAPI).editorManager?.editingEditor) {
@@ -515,6 +526,11 @@ export function bindContainerDomListener(eventManager: EventManager) {
     table.eventManager.isDown = false;
     table.eventManager.isDraging = false;
     table.eventManager.inertiaScroll.endInertia();
+    if (stateManager.isResizeCol()) {
+      endResizeCol(table);
+    } else if (stateManager.isResizeRow()) {
+      endResizeRow(table);
+    }
   };
   eventManager.globalEventListeners.push({
     name: 'pointerup',
@@ -563,7 +579,12 @@ export function bindContainerDomListener(eventManager: EventManager) {
     }
     const isSelecting = table.stateManager.isSelecting();
 
-    if (eventManager.isDraging && isSelecting && table.stateManager.select.ranges?.length > 0) {
+    if (
+      eventManager._enableTableScroll &&
+      eventManager.isDraging &&
+      isSelecting &&
+      table.stateManager.select.ranges?.length > 0
+    ) {
       // 检测鼠标是否离开了table
       const drawRange = table.getDrawRange();
       // const element = table.getElement();