+++ packages/vtable/src/scenegraph/group-creater/cell-type/chart-cell.ts
@@ -123,3 +123,5 @@ export function createChartCellGroup(
 
   return cellGroup;
 }
+
+export type CreateChartCellGroup = typeof createChartCellGroup;