+++ packages/vtable/src/scenegraph/layout/compute-col-width.ts
@@ -14,8 +14,8 @@ import { getQuadProps } from '../utils/padding';
 import { getProp } from '../utils/get-prop';
 import type { BaseTableAPI, HeaderData } from '../../ts-types/base-table';
 import type { PivotHeaderLayoutMap } from '../../layout/pivot-header-layout';
-import { getAxisConfigInPivotChart } from '../../layout/chart-helper/get-axis-config';
-import { computeAxisComponentWidth } from '../../components/axis/get-axis-component-size';
+import type { ComputeAxisComponentWidth } from '../../components/axis/get-axis-component-size';
+import { Factory } from '../../core/factory';
 import { Group as VGroup } from '@src/vrender';
 import { isArray, isFunction, isNumber, isObject, isValid } from '@visactor/vutils';
 import { decodeReactDom, dealPercentCalc } from '../component/custom';
@@ -33,6 +33,12 @@ export function computeColsWidth(table: BaseTableAPI, colStart?: number, colEnd?
     //   for (let col = colStart; col <= colEnd; col++) {
     //     table._clearColRangeWidthsMap(col);
     //   }
+  }
+
+  const layoutMap = table.internalProps.layoutMap;
+  if (table.isPivotTable()) {
+    (layoutMap as PivotHeaderLayoutMap).enableUseGetBodyCache();
+    (layoutMap as PivotHeaderLayoutMap).enableUseHeaderPathCache();
   }
 
   const oldColWidths: number[] = [];
@@ -95,7 +101,7 @@ export function computeColsWidth(table: BaseTableAPI, colStart?: number, colEnd?
     let endCol = table.colCount;
     if (table.widthAdaptiveMode === 'only-body') {
       for (let col = 0; col < table.colCount; col++) {
-        const colWidth = update ? newWidths[col] : table.getColWidth(col);
+        const colWidth = update ? newWidths[col] ?? table.getColWidth(col) : table.getColWidth(col);
         if (
           col < table.rowHeaderLevelCount ||
           (table.isPivotChart() && col >= table.colCount - table.rightFrozenColCount)
@@ -146,7 +152,7 @@ export function computeColsWidth(table: BaseTableAPI, colStart?: number, colEnd?
     let actualHeaderWidth = 0;
     let actualWidth = 0;
     for (let col = 0; col < table.colCount; col++) {
-      const colWidth = update ? newWidths[col] : table.getColWidth(col);
+      const colWidth = update ? newWidths[col] ?? table.getColWidth(col) : table.getColWidth(col);
       if (
         col < table.rowHeaderLevelCount ||
         (table.isPivotChart() && col >= table.colCount - table.rightFrozenColCount)
@@ -198,7 +204,7 @@ export function computeColsWidth(table: BaseTableAPI, colStart?: number, colEnd?
     for (let col = 0; col < table.colCount; col++) {
       // newColWidth could not be in column min max range possibly
       // const newColWidth = table._adjustColWidth(col, newWidths[col]) ?? table.getColWidth(col);
-      const newColWidth = newWidths[col] ?? table.getColWidth(col);
+      const newColWidth = newWidths[col] ?? table.getColWidth(col) ?? table.getColWidth(col);
       if (newColWidth !== oldColWidths[col]) {
         // update the column width in scenegraph
         table._setColWidth(col, newColWidth, false, true);
@@ -219,6 +225,11 @@ export function computeColsWidth(table: BaseTableAPI, colStart?: number, colEnd?
     table.scenegraph.updateContainer(true);
   }
   // console.log('computeColsWidth  time:', (typeof window !== 'undefined' ? window.performance.now() : 0) - time, colStart, colEnd);
+
+  if (table.isPivotTable()) {
+    (layoutMap as PivotHeaderLayoutMap).disableUseGetBodyCache();
+    (layoutMap as PivotHeaderLayoutMap).disableUseHeaderPathCache();
+  }
 }
 
 /**
@@ -301,8 +312,9 @@ function computeAutoColWidth(
     // 判断透视图轴组件
     if (table.isPivotChart()) {
       const layout = table.internalProps.layoutMap as PivotHeaderLayoutMap;
-      const axisConfig = getAxisConfigInPivotChart(col, row, layout);
+      const axisConfig = layout.getAxisConfigInPivotChart(col, row);
       if (axisConfig) {
+        const computeAxisComponentWidth: ComputeAxisComponentWidth = Factory.getFunction('computeAxisComponentWidth');
         const axisWidth = computeAxisComponentWidth(axisConfig, table);
         if (typeof axisWidth === 'number') {
           maxWidth = Math.max(axisWidth, maxWidth);
@@ -373,7 +385,7 @@ function computeAutoColWidth(
           cellHierarchyIndent += table.internalProps.headerHelper.getHierarchyIconWidth();
         }
       }
-    } else {
+    } else if (table.isListTable()) {
       deltaRow = prepareDeltaRow;
       // 基本表格表身body单元格 如果是树形展开 需要考虑缩进值
       // const cellHierarchyState = table.getHierarchyState(col, row);
@@ -594,7 +606,7 @@ function computeTextWidth(col: number, row: number, cellType: ColumnTypeOption, 
     text = cellValue;
   }
   const lines = breakString(text, table).text;
-  if (lines.length >= 1) {
+  if (lines.length >= 1 && !(lines.length === 1 && lines[0] === '')) {
     // eslint-disable-next-line no-loop-func
     lines.forEach((line: string) => {
       const width = table.measureText(line, {
@@ -695,7 +707,7 @@ export function getAdaptiveWidth(
   const sparklineColumns = [];
   let totalSparklineAbleWidth = 0;
   for (let col = startCol; col < endColPlus1; col++) {
-    const width = update ? newWidths[col] : table.getColWidth(col);
+    const width = update ? newWidths[col] ?? table.getColWidth(col) : table.getColWidth(col);
     const maxWidth = table.getMaxColWidth(col);
     const minWidth = table.getMinColWidth(col);
     if (width !== maxWidth && width !== minWidth) {
@@ -747,12 +759,12 @@ export function getAdaptiveWidth(
         totalDrawWidth -
         adaptiveColumns.reduce((acr, cur, index) => {
           if (cur !== col) {
-            return acr + (update ? newWidths[cur] : table.getColWidth(cur));
+            return acr + (update ? newWidths[cur] ?? table.getColWidth(col) : table.getColWidth(cur));
           }
           return acr;
         }, 0);
     } else {
-      colWidth = Math.round((update ? newWidths[col] : table.getColWidth(col)) * factor);
+      colWidth = Math.round((update ? newWidths[col] ?? table.getColWidth(col) : table.getColWidth(col)) * factor);
     }
     if (update) {
       newWidths[col] = table._adjustColWidth(col, colWidth);