+++ packages/vtable/src/scenegraph/scenegraph.ts
@@ -72,7 +72,7 @@ import { dealWithAnimationAppear } from './animation/appear';
 registerForVrender();
 
 // VChart poptip theme
-loadPoptip();
+// loadPoptip();
 container.load(splitModule);
 container.load(textMeasureModule);
 // container.load(renderServiceModule);
@@ -808,12 +808,41 @@ export class Scenegraph {
    * recalculates column width in all autowidth columns
    */
   recalculateColWidths() {
-    computeColsWidth(this.table, 0, this.table.colCount - 1, true);
+    const table = this.table;
+
+    if (table.widthMode === 'adaptive' || table.autoFillWidth || table.internalProps.transpose) {
+      computeColsWidth(this.table, 0, this.table.colCount - 1, true);
+    } else {
+      table._clearColRangeWidthsMap();
+      // left frozen
+      if (table.frozenColCount > 0) {
+        computeColsWidth(this.table, 0, table.frozenColCount - 1, true);
+      }
+      // right frozen
+      if (table.rightFrozenColCount > 0) {
+        computeColsWidth(this.table, table.rightFrozenColCount, table.colCount - 1, true);
+      }
+      // body
+      computeColsWidth(table, this.proxy.colStart, this.proxy.colEnd, true);
+    }
   }
 
   recalculateRowHeights() {
-    this.table.internalProps.useOneRowHeightFillAll = false;
-    computeRowsHeight(this.table, 0, this.table.rowCount - 1, true, true);
+    const table = this.table;
+    table.internalProps.useOneRowHeightFillAll = false;
+    if (table.heightMode === 'adaptive' || table.autoFillHeight) {
+      computeRowsHeight(this.table, 0, this.table.rowCount - 1, true, true);
+    } else {
+      // top frozen
+      if (table.frozenRowCount > 0) {
+        computeRowsHeight(this.table, 0, table.frozenRowCount - 1, true, true);
+      }
+      // bottom frozen
+      if (table.bottomFrozenRowCount > 0) {
+        computeRowsHeight(this.table, table.bottomFrozenRowCount, table.rowCount - 1, true, true);
+      }
+      computeRowsHeight(table, this.proxy.rowStart, this.proxy.rowEnd, true, true);
+    }
   }
 
   resize() {
@@ -1619,7 +1648,7 @@ export class Scenegraph {
       const drawRange = this.table.getDrawRange();
       if (abstractY >= drawRange.top && abstractY <= drawRange.bottom) {
         // to do: 处理最后一列外调整列宽
-        cell = this.table.getCellAt(abstractX - offset, abstractY);
+        cell = this.table.getCellAtRelativePosition(abstractX - offset, abstractY);
         return cell;
       }
       return { col: -1, row: -1 };