+++ packages/vtable/src/ts-types/table-engine.ts
@@ -21,8 +21,10 @@ import type {
   IRowDimension,
   ITitleDefine
 } from './pivot-table';
+
 import type { ColumnDefine, ColumnsDefine } from './list-table';
 import type { ICellAxisOption, ITableAxisOption } from './component/axis';
+
 import type { IEditor } from '@visactor/vtable-editors';
 import type { ITextStyleOption } from '../body-helper/style';
 import type { DataSource } from '../data';
@@ -273,8 +275,14 @@ export interface ListTableAPI extends BaseTableAPI {
   //#region 编辑器相关demo
   /** 获取单元格配置的编辑器 */
   getEditor: (col: number, row: number) => IEditor;
-  /** 开启单元格编辑 */
-  startEditCell: (col?: number, row?: number) => void;
+  /**
+   * 开启单元格编辑
+   * @param col
+   * @param row
+   * @param value 如果想要改变显示到编辑框中的值 可以value来设置改变
+   * @returns
+   */
+  startEditCell: (col?: number, row?: number, value?: string | number) => void;
   /** 结束编辑 */
   completeEditCell: () => void;
   //#endregion