+++ packages/vtable/src/vutil-extension-temp/spec/index.ts
@@ -1,3 +1,0 @@
-export * from './clone-deep';
-export * from './common';
-export * from './merge-spec';