+++ packages/vtable/src/vutil-extension-temp/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './object';
-export * from './polar';
-export * from './text';